requirements.txt
captum==0.5.0
graphviz==0.20.1
Markdown==3.4.1
matplotlib==3.6.2
numpy==1.22.3
opencv_python_headless==4.6.0.66
pandas==1.5.2
Pillow==9.4.0
plotly==5.11.0
scipy==1.10.1
setuptools==65.5.0
streamlit==1.19.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
transformers==4.25.1
git+https://github.com/vlue-c/Visual-Explanation-Methods-PyTorch.git